% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.R
\name{add_opt_cache}
\alias{add_opt_cache}
\title{Add component opt cache}
\usage{
add_opt_cache(cfg = list())
}
\arguments{
\item{cfg}{List of configuration options}
}
\description{
Adds a "cache" and "signature" attribute to the list of
configuration options.  The cache stores pre-computed component opt lists.
Once done this significantly speeds up the drawing of piecepack components
with that configuration list.  If you later change the configuration list you
should run this again to re-compute the cache.
}
\examples{
 \donttest{
   cfg <- list()
   system.time(replicate(500, draw_component("tile_face", cfg, 4, 4)))
   system.time(cfg <- add_opt_cache(cfg))
   system.time(replicate(500, draw_component("tile_face", cfg, 4, 4)))
 }
  
}
