% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{c2o}
\alias{c2o}
\title{Read piecepack configuration list from command-line arguments}
\usage{
c2o(args = NULL)
}
\arguments{
\item{args}{Character vector of command-line arguments.}
}
\description{
Read piecepack configuration list from command-line arguments
}
