% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{configure}
\alias{configure}
\title{Generate piecepack configuration JSON}
\usage{
configure(args = commandArgs(TRUE))
}
\arguments{
\item{args}{Character vector of command-line arguments.}
}
\description{
Processes command-line options and writes to JSON.
}
