% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.r
\name{configure_piecepack}
\alias{configure_piecepack}
\title{Generate piecepack configuration JSON}
\usage{
configure_piecepack(args = commandArgs(TRUE))
}
\arguments{
\item{args}{Character vector of command-line arguments.}
}
\description{
Processes command-line options and writes to JSON.
}
