% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_draw_fn.R
\name{draw_fn_helpers}
\alias{draw_fn_helpers}
\alias{add_gridlines}
\alias{add_mat}
\alias{add_checkers}
\alias{add_hexlines}
\alias{get_grid_shape}
\alias{is_color_invisible}
\title{\code{basic_draw_fn} helper functions}
\usage{
add_gridlines(col, shape = "rect", shape_t = 90)

add_mat(col, shape = "rect", shape_t = 90, mat_width = 0)

add_checkers(col, shape = "rect", shape_t = 90)

add_hexlines(col, shape = "rect")

get_grid_shape(shape, shape_t = 90, shape_r = 0.2)

is_color_invisible(col)
}
\arguments{
\item{col}{Color}

\item{shape}{String of the shape}

\item{shape_t}{Angle (in degrees) of first vertex of shape (ignored by many shapes).}

\item{mat_width}{Numeric vector of mat widths}

\item{shape_r}{Radial distance (from 0 to 0.5) (ignored by most shapes)}
}
\description{
\code{add_gridlines} adds gridlines.
\code{add_mat} adds a mat.
\code{add_checkers} adds checkers.
\code{add_hexlines} adds hexlines.
\code{get_grid_shape} returns a function that draws the component shape.
\code{is_color_invisible} tells whether the color is transparent (and hence need not be drawn).
}
