% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.R
\name{draw_preview}
\alias{draw_preview}
\title{Draw piecepack deck preview}
\usage{
draw_preview(cfg = list())
}
\arguments{
\item{cfg}{Piecepack configuration list}
}
\description{
Draw piecepack deck preview
}
