% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_systems.R
\name{game_systems}
\alias{game_systems}
\title{Standard game systems}
\usage{
game_systems(style = NULL)
}
\arguments{
\item{style}{If \code{NULL} (the default) uses suit glyphs from the default \dQuote{sans} font.
If \code{"dejavu"} it will use suit glyphs from the "DejaVu Sans" font (must be installed on the system).}
}
\description{
\code{game_systems} returns a list of \code{pp_cfg} objects
representing several game systems.
}
\details{
Contains the following game systems:\describe{
\item{dual_piecepacks_expansion}{A companion piecepack with a special suit scheme.
              See \url{https://trevorldavis.com/piecepackr/dual-piecepacks-pnp.html}.}
\item{hexpack}{A hexagonal extrapolation of the piecepack designed by Nathan Morse and Daniel Wilcox.
               See \url{https://boardgamegeek.com/boardgameexpansion/35424/hexpack}.}
\item{icehouse_pieces}{Icehouse pieces aka Looney Pyramids is a game system invented by Andrew Looney.
                       See \url{https://www.looneylabs.com/looney-pyramids}.}
\item{piecepack}{A public domain game system invented by James "Kyle" Droscha.
  See \url{http://www.ludism.org/ppwiki}.
  Configuration also contains the following piecepack accessories:\describe{
    \item{piecepack matchsticks}{A public domain accessory developed by Dan Burkey.
                                See \url{http://www.ludism.org/ppwiki/PiecepackMatchsticks}.}
    \item{piecepack pyramids}{A public domain accessory developed by Tim Schutz.
                             See \url{http://www.ludism.org/ppwiki/PiecepackPyramids}.}
    \item{piecepack saucers}{A public domain accessory developed by Karol M. Boyle at Mesomorph Games.
                             See \url{http://www.piecepack.org/Accessories.html}.}
  }}
\item{playing_cards_expansion}{A piecepack with the standard (``French'') playing card suits.
                               See \url{http://www.ludism.org/ppwiki/PlayingCardsExpansion}.}
\item{subpack}{A mini piecepack.  Designed to be used with the \code{piecepack} to make piecepack
              ``stackpack'' diagrams.  See \url{http://www.ludism.org/ppwiki/StackPack}.}
}
}
\examples{
       cfgs <- game_systems()
       names(cfgs)

       df_tiles <- data.frame(piece_side="tile_back", x=0.5+c(3,1,3,1), y=0.5+c(3,3,1,1),
                              suit=NA, angle=NA, z=NA, stringsAsFactors=FALSE)
       df_coins <- data.frame(piece_side="coin_back", x=rep(4:1, 4), y=rep(4:1, each=4),
                              suit=c(1,4,1,4,4,1,4,1,2,3,2,3,3,2,3,2),
                              angle=rep(c(180,0), each=8), z=1/4+1/16, stringsAsFactors=FALSE)
       df <- rbind(df_tiles, df_coins)
       df$cfg <- "playing_cards_expansion"

       pmap_piece(df, envir=cfgs, op_scale=0.5, default.units="in")
}
\seealso{
\code{\link{pp_cfg}} for information about the \code{pp_cfg} objects returned by \code{game_systems}.
}
