% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_arrangement_opts}
\alias{get_arrangement_opts}
\title{Get collect piecepacks options}
\usage{
get_arrangement_opts(args = commandArgs(TRUE))
}
\arguments{
\item{args}{Character vector of command-line arguments.}
}
\description{
Get collect piecepacks options.
}
