% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_embedded_font}
\alias{get_embedded_font}
\title{Get embedded font utility function}
\usage{
get_embedded_font(font, char)
}
\arguments{
\item{font}{A character vector of font(s) passed to the \code{fontfamily} argument of \code{grid::gpar}.}

\item{char}{A character vector of character(s) to be embedded by \code{grid::grid.text}}
}
\value{
A character vector of fonts that were actually embedded by \code{cairo_pdf}.  \code{NA}'s means no embedded font detected. 
       This either means that no font was found or that a color emoji font was found and instead of a font an image was embedded.
}
\description{
This (vectorized) function returns which font is actually embedded by \code{cairo_pdf}.
}
\details{
This functions depends on \code{pdffonts} being on the system path.
}
