% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_fns.R
\name{grid_shape_grobs}
\alias{grid_shape_grobs}
\alias{halmaGrob}
\alias{pyramidGrob}
\alias{kiteGrob}
\alias{convexGrobFn}
\alias{concaveGrobFn}
\title{Grid shape grob utility functions}
\usage{
halmaGrob(gp = gpar(), name = NULL)

pyramidGrob(gp = gpar(), name = NULL)

kiteGrob(gp = gpar(), name = NULL)

convexGrobFn(n_vertices, t)

concaveGrobFn(n_vertices, t, r = 0.2)
}
\arguments{
\item{gp}{An object of class ‘gpar’, typically the output from a call
to the function ‘gpar’.  This is basically a list of
graphical parameter settings.}

\item{name}{A character identifier (for grid)}

\item{n_vertices}{Number of vertices}

\item{t}{Angle (in degrees) of first vertex of shape}

\item{r}{Radial distance (from 0 to 0.5)}
}
\description{
Utility functions that produce grobs of various shapes 
or function that returns a function that produces a grob.
These are usually wrappers of \code{polygonGrob} or \code{pathGrob}.
}
