% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_functions.R
\name{inch}
\alias{inch}
\title{Inch utility function}
\usage{
inch(x)
}
\arguments{
\item{x}{Number representing number of inches}
}
\description{
This utility function is equivalent to \code{grid::unit(x, "in")}.
}
