% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{load_configurations}
\alias{load_configurations}
\title{Load piecepack configurations}
\usage{
load_configurations(cfg = list(), cfg_files = character(0),
  internal_cfgs = character(0))
}
\arguments{
\item{cfg}{Configuration list to augment}

\item{cfg_files}{Filenames of json configurations.}

\item{internal_cfgs}{Strings of internal piecepackr configurations.}
}
\description{
Load in piecepack configurations
}
