% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_functions.R
\name{make_collection}
\alias{make_collection}
\title{Collects print-and-play piecepack pdfs into one pdf}
\usage{
make_collection(output_filename, input_filenames, size = "letter",
  metadata = list())
}
\arguments{
\item{output_filename}{Filename of PnP output}

\item{input_filenames}{Vector of input filenames (starting with preview pdf)}

\item{size}{PnP output size (currently either "letter" or "A4")}

\item{metadata}{List of metadata to embed}
}
\description{
Collects print-and-play piecepack pdfs into one pdf
with a preview at the beginning.
}
