% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_functions.R
\name{make_collection_preview}
\alias{make_collection_preview}
\title{Make collection preview}
\usage{
make_collection_preview(cfg)
}
\arguments{
\item{cfg}{Piecepack configuration list}
}
\description{
Makes a preview pdf of a collection of pnp decks.
}
