% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.R
\name{make_images}
\alias{make_images}
\title{Make piecepack images}
\usage{
make_images(cfg = list(), directory = tempdir(), format = "svg",
  angles = 0)
}
\arguments{
\item{cfg}{Piecepack configuration list}

\item{directory}{Directory where to place images}

\item{format}{Format}

\item{angle}{Angle to rotate images (in degrees)}
}
\description{
Makes images of individual piecepack components.
}
