% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.r
\name{make_images_schemes}
\alias{make_images_schemes}
\title{Make images}
\usage{
make_images_schemes(arg, ...)
}
\arguments{
\item{arg}{Arg}
}
\value{
\code{invisible(NULL)}
}
\description{
This function makes image schemes
}
