% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.R
\name{make_piecepack_images}
\alias{make_piecepack_images}
\title{Make piecepack images}
\usage{
make_piecepack_images(cfg)
}
\arguments{
\item{cfg}{Piecepack configuration list}
}
\description{
Makes images of individual piecepack components.
}
