% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.R
\name{make_piecepack_preview}
\alias{make_piecepack_preview}
\title{Make piecepack deck preview svg}
\usage{
make_piecepack_preview(cfg = c2o())
}
\arguments{
\item{cfg}{Piecepack configuration list}
}
\description{
Make piecepack deck preview svg
}
