% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_functions.r
\name{make_pnp_piecepack}
\alias{make_pnp_piecepack}
\title{Make print-and-play piecepack pdf}
\usage{
make_pnp_piecepack(cfg)
}
\arguments{
\item{cfg}{Piecepack configuration list}
}
\description{
Makes a print-and-play piecepack pdf.
}
