% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.R
\name{make_preview}
\alias{make_preview}
\title{Make piecepack deck preview svg}
\usage{
make_preview(cfg = list(),
  output_filename = "svg/previews/piecepack_deck.svg")
}
\arguments{
\item{cfg}{Piecepack configuration list}

\item{output_filename}{Filename of PnP output}
}
\description{
Make piecepack deck preview svg
}
