% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentGrobFns.R
\name{pieceGrobFns}
\alias{pieceGrobFns}
\alias{basicPieceGrobFn}
\alias{pyramidTopGrob}
\title{Piece Grob Functions}
\usage{
basicPieceGrobFn(piece_side, i_s, i_r, cfg)

pyramidTopGrob(piece_side, i_s, i_r, cfg)
}
\arguments{
\item{piece_side}{A string with piece and side separated by a underscore e.g. "coin_face"}

\item{i_s}{Number of suit}

\item{i_r}{Number of rank}

\item{cfg}{Piecepack configuration list or \code{pp_cfg} object, 
a list of \code{pp_cfg} objects,
or a character vector of \code{pp_cfg} objects}
}
\description{
Default functions \code{grid.piece} uses to create \code{grid} 
graphical \code{grob} objects.
}
