% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepackr-defunct.R
\name{piecepackr-defunct}
\alias{piecepackr-defunct}
\alias{halmaGrob}
\alias{kiteGrob}
\alias{pyramidGrob}
\alias{convexGrobFn}
\alias{concaveGrobFn}
\alias{gridlinesGrob}
\alias{matGrob}
\alias{checkersGrob}
\alias{hexlinesGrob}
\alias{get_shape_grob_fn}
\title{Defunct functions}
\usage{
halmaGrob(...)

kiteGrob(...)

pyramidGrob(...)

convexGrobFn(...)

concaveGrobFn(...)

gridlinesGrob(...)

matGrob(...)

checkersGrob(...)

hexlinesGrob(...)

get_shape_grob_fn(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
These functions are Defunct and have been removed from piecepackr.
}
\details{
\enumerate{
\item{For \code{get_shape_grob_fn} use \code{pp_shape()$shape} instead.}
\item{For \code{gridlinesGrob()} use \code{pp_shape()$gridlines()} instead.}
\item{For \code{matGrob()} use \code{pp_shape()$mat()} instead.}
\item{For \code{checkersGrob()}() use \code{pp_shape()$checkers()} instead.}
\item{For \code{hexlinesGrob()} use \code{pp_shape()$hexlines()} instead.}
\item{For \code{halmaGrob()} use \code{pp_shape("halma")$shape()} instead.}
\item{For \code{kiteGrob()} use \code{pp_shape("kite")$shape()} instead.}
\item{For \code{pyramidGrob()} use \code{pp_shape("pyramid")$shape()} instead.}
\item{For \code{convexGrobFn(n, t)} use \code{pp_shape(paste0("convex", n), t)$shape} instead.}
\item{For \code{concaveGrobFn(n, t, r)} use \code{pp_shape(paste0("concave", n), t, r)$shape} instead.}
}
}
