% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{read_configuration}
\alias{read_configuration}
\title{Read piecepack configuration JSON}
\usage{
read_configuration(cfg_file = "stdin")
}
\arguments{
\item{cfg_file}{Filename of json configuration.  Default is to read in from standard input.}
}
\description{
Read in and process piecepack configuration JSON.
}
