% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_piece_obj.R
\name{save_piece_obj}
\alias{save_piece_obj}
\title{Save Wavefront OBJ files of board game pieces}
\usage{
save_piece_obj(
  piece_side = "tile_face",
  suit = 1,
  rank = 1,
  cfg = pp_cfg(),
  ...,
  x = 0,
  y = 0,
  z = 0,
  angle = 0,
  axis_x = 0,
  axis_y = 0,
  width = NA,
  height = NA,
  depth = NA,
  filename = tempfile(fileext = ".obj"),
  scale = 1,
  res = 72
)
}
\arguments{
\item{piece_side}{A string with piece and side separated by a underscore e.g. "coin_face"}

\item{suit}{Number of suit (starting from 1).}

\item{rank}{Number of rank (starting from 1)}

\item{cfg}{Piecepack configuration list or \code{pp_cfg} object,
a list of \code{pp_cfg} objects,
or a character vector of \code{pp_cfg} objects}

\item{...}{Ignored.}

\item{x}{Where to place piece on x axis of viewport}

\item{y}{Where to place piece on y axis of viewport}

\item{z}{z-coordinate of the piece.  Has no effect if \code{op_scale} is \code{0}.}

\item{angle}{Angle (on xy plane) to draw piece at}

\item{axis_x}{First coordinate of the axis unit vector.}

\item{axis_y}{Second coordinate of the axis unit vector.}

\item{width}{Width of piece}

\item{height}{Height of piece}

\item{depth}{Depth (thickness) of piece.  Has no effect if \code{op_scale} is \code{0}.}

\item{filename}{Name of Wavefront OBJ object.}

\item{scale}{Multiplicative scaling factor to apply to width, height, and depth.}

\item{res}{Resolution of the faces.}
}
\value{
A list with named elements "obj", "mtl", "png" with the created filenames.
}
\description{
\code{save_piece_obj} saves Wavefront OBJ files (including associated MTL and texture image) of board game pieces.
}
\examples{
    cfg <- game_systems("dejavu3d")$piecepack
    files <- save_piece_obj("tile_face", suit = 3, rank = 3, cfg = cfg)
    print(files)
}
\seealso{
See \code{\link{geometry_utils}} for a discussion of the 3D rotation parameterization.
}
