% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.R
\name{split}
\alias{split}
\title{Convert delimiter separated string to vector}
\usage{
split(x, sep = ",", float = FALSE, color = FALSE)
}
\arguments{
\item{x}{String to convert}

\item{sep}{Delimiter (defaults to ",")}

\item{float}{If `TRUE` cast to numeric}

\item{color}{if `TRUE` convert empty strings to `"transparent"`}
}
\description{
Converts delimiter separated string to a vector.
}
