% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.r
\name{draw_component}
\alias{draw_component}
\title{Draw piecepack component}
\usage{
draw_component(component_side, cfg = c2o(), i_s = cfg$i_unsuit,
  i_r = 1, x = 0.5, y = 0.5, width = NULL, height = NULL,
  svg = FALSE, ...)
}
\arguments{
\item{component_side}{A string with component and side separated by a underscore e.g. "coin_face"}

\item{cfg}{Piecepack configuration list}

\item{i_s}{Number of suit}

\item{i_r}{Number of rank}

\item{x}{Number between 0 and 1 of where to place component}

\item{y}{Number between 0 and 1 of where to place component}

\item{width}{Width of component}

\item{height}{Height of component}

\item{svg}{If \code{TRUE} instead of drawing directly into graphics device
export to svg, re-import svg, and then draw it to graphics device.  
This is useful if drawing really big or small and don't want
to play with re-configuring fontsizes.}

\item{...}{Extra arguments to pass to \code{grid::viewport} like \code{angle}.}
}
\description{
Draws a piecepack component onto the graphics device
}
