% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieceGrob.R
\name{grid.piece}
\alias{grid.piece}
\alias{pmap_piece}
\alias{pieceGrob}
\title{Draw piecepack pieces using grid}
\usage{
pmap_piece(.l, ..., draw = TRUE, name = NULL, gp = NULL, vp = NULL)

pieceGrob(piece_side = "tile_back", suit = NA, rank = NA,
  cfg = pp_cfg(), x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  angle = 0, use_pictureGrob = FALSE, width = NA, height = NA,
  default.units = "npc", envir = NULL, name = NULL, gp = NULL,
  vp = NULL)

grid.piece(piece_side = "tile_back", suit = NA, rank = NA,
  cfg = list(), x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  angle = 0, use_pictureGrob = FALSE, width = NA, height = NA,
  default.units = "npc", envir = NULL, name = NULL, gp = NULL,
  draw = TRUE, vp = NULL)
}
\arguments{
\item{.l}{A list of vectors, such as a data frame. The length of \code{.l}
determines the number of arguments that \code{grid.piece_wrapper}
will be called  with. List names will be used if present.}

\item{...}{Extra arguments to pass to \code{pieceGrob}.}

\item{draw}{A logical value indicating whether graphics output should be produced.}

\item{name}{A character identifier (for grid)}

\item{gp}{An object of class ‘gpar’, typically the output from a call
to the function ‘gpar’.  This is basically a list of
graphical parameter settings.}

\item{vp}{A \code{grid} viewport object (or NULL).}

\item{piece_side}{A string with piece and side separated by a underscore e.g. "coin_face"}

\item{suit}{Number of suit (highest rank starting from 1).
The number above the total number of suits is the neutral "unsuit".
and the next number above that is "no suits".}

\item{rank}{Number of rank (lowest rank starting from 1)}

\item{cfg}{Piecepack configuration list or \code{pp_cfg} object, 
a list of \code{pp_cfg} objects,
or a character vector of \code{pp_cfg} objects}

\item{x}{Where to place piece on x axis of viewport}

\item{y}{Where to place piece on y axis of viewport}

\item{angle}{Angle to draw piece at}

\item{use_pictureGrob}{If \code{TRUE} instead of directly returning the grob first 
export to (temporary) svg and then re-import as a \code{grImport2::pictureGrob}.  
This is useful if drawing pieces really big or small and don't want
to play with re-configuring fontsizes.}

\item{width}{Width of piece}

\item{height}{Height of piece}

\item{default.units}{A string indicating the default units to use if 
'x', 'y', 'width', and/or 'height' are only given as numeric vectors.}

\item{envir}{Environment (or named list) containing configuration list(s).}
}
\value{
A \code{grob} object.  If \code{draw} is \code{TRUE} then as a side effect
        will also draw it to the graphics device.
}
\description{
\code{grid.piece} draws a piecepack pieces onto the graphics device.  
\code{pieceGrob} is its \code{grid} \code{grob} counterpart.
\code{pmap_piece} operates on the rows of a data frame  
    applying \code{pieceGrob} to each row.
}
