% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_functions.R
\name{make_collection}
\alias{make_collection}
\title{Collects print-and-play piecepack pdfs into one pdf}
\usage{
make_collection(cfg)
}
\arguments{
\item{cfg}{Piecepack configuration list}
}
\description{
Collects print-and-play piecepack pdfs into one pdf
with a preview at the beginning.
}
