% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_functions.R
\name{make_collection_preview}
\alias{make_collection_preview}
\title{Make collection preview}
\usage{
make_collection_preview(output_filename, input_filenames, title = "",
  size = "letter")
}
\arguments{
\item{output_filename}{Filename of PnP output}

\item{input_filenames}{Vector of input filenames}

\item{title}{Title of collection}

\item{size}{PnP output size (currently either "letter" or "A4")}
}
\description{
Makes a preview pdf of a collection of pnp decks.
}
