% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_functions.R
\name{make_pnp}
\alias{make_pnp}
\title{Make print-and-play piecepack pdf}
\usage{
make_pnp(cfg, output_filename = "pdf/decks/piecepack_deck.pdf",
  size = "letter")
}
\arguments{
\item{cfg}{Piecepack configuration list}

\item{output_filename}{Filename of PnP output}

\item{size}{PnP output size (currently either "letter" or "A4")}
}
\description{
Makes a print-and-play piecepack pdf.
}
