% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentGrobFns.R
\name{pieceGrobFns}
\alias{pieceGrobFns}
\alias{basicPieceGrob}
\alias{pyramidTopGrob}
\title{Piece Grob Functions}
\usage{
basicPieceGrob(piece_side, suit, rank, cfg = pp_cfg())

pyramidTopGrob(piece_side, suit, rank, cfg = pp_cfg())
}
\arguments{
\item{piece_side}{A string with piece and side separated by a underscore e.g. "coin_face"}

\item{suit}{Number of suit (highest rank starting from 1).
The number above the total number of suits is the neutral "unsuit".
and the next number above that is "no suits".}

\item{rank}{Number of rank (lowest rank starting from 1)}

\item{cfg}{Piecepack configuration list or \code{pp_cfg} object,}
}
\description{
Default functions \code{grid.piece} uses to create \code{grid} 
graphical \code{grob} objects.
}
