% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepackr-package.R
\docType{package}
\name{piecepackr-package}
\alias{piecepackr}
\alias{piecepackr-package}
\title{piecepackr: Board Game Graphics}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Functions to make board game graphics.  Specializes in game diagrams, animations, and "Print & Play" layouts for the 'piecepack' <https://www.ludism.org/ppwiki> but can make graphics for other board game systems.  Includes configurations for several public domain game systems.
}
\section{Package options}{

The following piecepackr function arguments may be set globally via \code{\link[base:options]{base::options()}}:
\describe{
\item{piecepackr.cfg}{Sets a new default for the \code{cfg} argument}
\item{piecepackr.default.units}{Sets a new default for the \code{default.units} argument}
\item{piecepackr.envir}{Sets a new default for the \code{envir} argument}
\item{piecepackr.op_angle}{Sets a new default for the \code{op_angle} argument}
\item{piecepackr.op_scale}{Sets a new default for the \code{op_scale} argument}
\item{piecepackr.trans}{Sets a new default for the \code{trans} argument}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://trevorldavis.com/piecepackr/ (blog)}
  \item \url{https://trevorldavis.com/R/piecepackr (pkgdown)}
  \item \url{https://groups.google.com/forum/#!forum/piecepackr (forum)}
  \item Report bugs at \url{https://github.com/piecepackr/piecepackr/issues}
}

}
\author{
\strong{Maintainer}: Trevor L Davis \email{trevor.l.davis@gmail.com}

Other contributors:
\itemize{
  \item Delapouite <https://delapouite.com/> (Meeple shape extracted from "Meeple icon" <https://game-icons.net/1x1/delapouite/meeple.html> / "CC BY 3.0" <https://creativecommons.org/licenses/by/3.0/>) [data contributor]
}

}
\keyword{internal}
