% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecepackr-package.R
\docType{package}
\name{piecepackr-package}
\alias{piecepackr-package}
\alias{piecepackr}
\title{piecepackr: Board Game Graphics}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
Functions to make board game graphics.
Specializes in game diagrams, animations, and "Print & Play" layouts
for the 'piecepack' \url{https://www.ludism.org/ppwiki} but can make
graphics for other board game systems.
Includes configurations for several public domain game systems.
}
\section{Package options}{

The following piecepackr function arguments may be set globally via \code{\link[base:options]{base::options()}}:
\describe{
\item{piecepack.at.inform}{If \code{FALSE} turns off messages when affine transformation support not detected in active graphics device.}
\item{piecepackr.cfg}{Sets a new default for the \code{cfg} argument}
\item{piecepackr.default.units}{Sets a new default for the \code{default.units} argument}
\item{piecepackr.envir}{Sets a new default for the \code{envir} argument}
\item{piecepackr.op_angle}{Sets a new default for the \code{op_angle} argument}
\item{piecepackr.op_scale}{Sets a new default for the \code{op_scale} argument}
\item{piecepackr.trans}{Sets a new default for the \code{trans} argument}
}
}

\seealso{
Useful links:
\itemize{
\item blog: \url{https://trevorldavis.com/piecepackr/}
\item pkgdown: \url{https://trevorldavis.com/R/piecepackr/}
\item forum: \url{https://groups.google.com/forum/#!forum/piecepackr}
\item Report bugs: \url{https://github.com/piecepackr/piecepackr/issues}
}
}
