% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_images.R
\name{pp_cfg}
\alias{pp_cfg}
\title{Create piecepack configuration list with a cache of component opts}
\usage{
pp_cfg(cfg = list(), ...)
}
\arguments{
\item{cfg}{List of configuration options}
}
\description{
Adds a \code{cache} and \code{signature} attribute and a \code{pp_cfg} class
to a list of configuration options.  
The cache stores pre-computed component opt lists.
Once done this significantly speeds up the drawing of piecepack components
with that configuration list.  However if you later change the configuration list you
should run this again to re-compute the cache otherwise \code{draw_component} may 
not work as intended.
}
\examples{
 \donttest{
   cfg <- list()
   system.time(replicate(500, draw_component("tile_face", cfg, 4, 4)))
   system.time(cfg <- pp_cfg(cfg))
   system.time(replicate(500, draw_component("tile_face", cfg, 4, 4)))
 }
  
}
