% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_cfg.R
\name{pp_cfg}
\alias{pp_cfg}
\alias{is_pp_cfg}
\alias{as_pp_cfg}
\title{Configuration list R6 object}
\usage{
pp_cfg(cfg = list())

is_pp_cfg(cfg)

as_pp_cfg(cfg = list())
}
\arguments{
\item{cfg}{List of configuration options}
}
\description{
\code{pp_cfg} and \code{as_pp_cfg} creates piecepack configuration list R6 object.
\code{is_pp_cfg} returns \code{TRUE} if object is a piecepack configuration list R6 object.
\code{as.list} will convert it into a list.
}
\examples{
 \donttest{
   cfg <- list()
   system.time(replicate(500, grid.piece("tile_face", 4, 4, cfg)))
   cfg <- pp_cfg(list())
   system.time(replicate(500, grid.piece("tile_face", 4, 4, cfg)))
 }
  
}
