% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{read_configuration}
\alias{read_configuration}
\title{Read piecepack configuration JSON}
\usage{
read_configuration(args = commandArgs(TRUE), description = "")
}
\arguments{
\item{args}{Character vector of command-line arguments.}

\item{description}{Description to pass to \code{OptionParser}.}
}
\description{
Read in and process piecepack configuration JSON.
}
