% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnp_functions.R
\name{save_print_and_play}
\alias{save_print_and_play}
\title{Save piecepack print-and-play (PnP) file}
\usage{
save_print_and_play(cfg = list(), output_filename = "piecepack.pdf",
  size = "letter", pieces = c("piecepack", "matchsticks", "pyramids"))
}
\arguments{
\item{cfg}{Piecepack configuration list}

\item{output_filename}{Filename for print-and-play file}

\item{size}{PnP output size (currently either "letter", "A4", or "A5")}

\item{pieces}{Character vector of desired PnP pieces}
}
\description{
Save piecepack print-and-play (PnP) file
}
