% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_piece.R
\name{scale_x_piece}
\alias{scale_x_piece}
\alias{scale_y_piece}
\alias{label_letter}
\title{Alternative ggplot2 position scales}
\usage{
scale_x_piece(
  ...,
  name = NULL,
  breaks = positive_integers,
  minor_breaks = NULL
)

scale_y_piece(
  ...,
  name = NULL,
  breaks = positive_integers,
  minor_breaks = NULL
)

label_letter()
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} or \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}

\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
\link[scales:trans_new]{transformation object}
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output (e.g., a function returned by \code{\link[scales:breaks_extended]{scales::extended_breaks()}}).
Also accepts rlang \link[rlang:as_function]{lambda} function notation.
}}

\item{minor_breaks}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item \code{waiver()} for the default breaks (one minor break between
each major break)
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks. Also
accepts rlang \link[rlang:as_function]{lambda} function notation.
}}
}
\value{
\code{scale_x_piece()} and \code{scale_y_piece()} return ggplot2 scale objecs.
\code{label_letter()} returns a function suitable for use with a \code{labels} argument.
}
\description{
\code{scale_x_piece()} and \code{scale_y_piece()} are wrappers
around \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} and
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} with "better"
defaults for use with \code{\link[=geom_piece]{geom_piece()}}.
\code{label_letter()} labels numbers with letters
to more easily generate (i.e. chess) algebraic notation coordinates.
}
\examples{
if (require("ggplot2") && require("tibble")) {

  envir <- game_systems("sans")
  df_board <- tibble(piece_side = "board_face", suit = 3, rank = 8,
                 x = 4.5, y = 4.5)
  df_w <- tibble(piece_side = "bit_face", suit = 6, rank = 1,
                 x = rep(1:8, 2), y = rep(1:2, each=8))
  df_b <- tibble(piece_side = "bit_face", suit = 1, rank = 1,
                 x = rep(1:8, 2), y = rep(7:8, each=8))
  df <- rbind(df_board, df_w, df_b)
  # `cfg` must be a character vector for `geom_piece()`
  ggplot(df, aes_piece(df)) +
      geom_piece(cfg = "checkers1", envir = envir) +
      coord_fixed() +
      scale_x_piece(labels = label_letter()) +
      scale_y_piece() +
      theme_minimal(32) +
      theme(panel.grid = element_blank())
}
}
